+++ simpletransformers/classification/multi_modal_classification_model.py
@@ -38,7 +38,6 @@ from transformers.optimization import (
 from torch.optim import AdamW
 from transformers.optimization import Adafactor
 from transformers import (
-    BERT_PRETRAINED_MODEL_ARCHIVE_LIST,
     WEIGHTS_NAME,
     BertConfig,
     BertModel,
@@ -177,24 +176,6 @@ class MultiModalClassificationModel:
         self.model = MMBTForClassification(
             self.config, self.transformer, self.img_encoder
         )
-
-        if model_name not in BERT_PRETRAINED_MODEL_ARCHIVE_LIST:
-            try:
-                self.model.load_state_dict(
-                    torch.load(os.path.join(model_name, "pytorch_model.bin"))
-                )
-            except EnvironmentError:
-                msg = (
-                    "Model name '{}' was not found in model name list ({}). "
-                    "We assumed '{}' was a path or url to model weight files named one of {} but "
-                    "couldn't find any such file at this path or url.".format(
-                        model_name,
-                        ", ".join(BERT_PRETRAINED_MODEL_ARCHIVE_LIST),
-                        model_name,
-                        "pytorch_model.bin",
-                    )
-                )
-                raise EnvironmentError(msg)
 
         self.tokenizer = tokenizer_class.from_pretrained(
             model_name, do_lower_case=self.args.do_lower_case, **kwargs